* **29.08.2019** - continuing **implementing different rarefaction depths analysis**   * adjusted script `/Users/paul/Documents/CU_combined/Github/150_calculate_fasttree.sh`     * added check for readily available data - ran ok   * adjusted script `/Users/paul/Documents/CU_combined/Github/155_filter_data_to_match_trees.sh`     * added check for readily data - ran ok   * adjusted script `~/Documents/CU_combined/Github/160_alpha_rarefaction_curves_phylogenetic.sh`     * adjusted case and added check for readily available data     * not run, no new insights gained - available available via old plot     * also did not run `~/Documents/CU_combined/Github/125_alpha_rarefaction_curves_non_phylogenetic.sh` - results available via old plot     * commit with further comments in commit message - rarefaction plots not re-run - do need to re-run both summary plots - mainly needed are abort checks if data is already available

+++ 150_calculate_fasttree.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-# 01.06.2019 - Paul Czechowski - paul.czechowski@gmail.com 
+# 29.08.2019 - Paul Czechowski - paul.czechowski@gmail.com 
 # ========================================================
 # Calculating trees from masked alignments using FastTree - which uses multiple cores well.
 
@@ -118,16 +118,25 @@ done
 
 for k in "${!inpth_seq[@]}"; do
   
-  printf "\n${bold}$(date):${normal} Calculating tree ${inpth_seq[$k]}...\n"
-  qiime phylogeny fasttree \
-    --i-alignment "${inpth_seq[$k]}" \
-    --p-n-threads "$thrds" \
-    --o-tree "${otpth_tree[$k]}" \
-    --verbose  2>&1 | tee -a "${otpth_log[$k]}"
+  if [ ! -f "${otpth_tree[$k]}" ]; then
   
-  printf "\n${bold}$(date):${normal} Midpoint-rooting "${otpth_tree[$k]}"...\n"  
-  qiime phylogeny midpoint-root \
-    --i-tree "${otpth_tree[$k]}" \
-    --o-rooted-tree "${otpth_rtree[$k]}"
+    printf "\n${bold}$(date):${normal} Calculating tree ${inpth_seq[$k]}...\n"
+    qiime phylogeny fasttree \
+      --i-alignment "${inpth_seq[$k]}" \
+      --p-n-threads "$thrds" \
+      --o-tree "${otpth_tree[$k]}" \
+      --verbose  2>&1 | tee -a "${otpth_log[$k]}"
+  
+    printf "\n${bold}$(date):${normal} Midpoint-rooting "${otpth_tree[$k]}"...\n"  
+    qiime phylogeny midpoint-root \
+      --i-tree "${otpth_tree[$k]}" \
+      --o-rooted-tree "${otpth_rtree[$k]}"
+  
+  else
+ 
+    # diagnostic message
+    printf "${bold}$(date):${normal} Analysis already done for \"$(basename "${inpth_seq[$k]}")\"...\n"
+
+  fi
 
 done
